requirements.txt
absl-py==2.1.0
accelerate==0.20.3
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==43.0.1
cycler==0.11.0
datasets==2.13.2
dill==0.3.6
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mmdet==2.17.0
mmengine==0.8.1
mmsegmentation==0.19.0
model-index==0.1.11
multidict==6.0.5
multiprocess==0.70.14
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
#pandas==1.1.5
Pillow==9.5.0
pip==22.3.1
platformdirs==4.0.0
prettytable==3.7.0
protobuf==3.20.3
psutil==6.1.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.21.0
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.1
regex==2024.4.16
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scipy==1.7.3
setuptools==60.2.0
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
timm==0.9.12
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.2
transformers==4.30.2
ttach==0.0.3
typing_extensions==4.7.1
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.38.4
xxhash==3.5.0
yapf==0.40.2
yarl==1.9.4
zipp==3.15.0
